+++ sdk/eventhub/Microsoft.Azure.WebJobs.Extensions.EventHubs/src/Config/EventHubOptions.cs
@@ -181,8 +181,8 @@ namespace Microsoft.Azure.WebJobs.EventHubs
         private JObject ConstructInitialOffsetOptions() =>
             new JObject
                 {
-                    { nameof(InitialOffsetOptions.Type), InitialOffsetOptions.Type },
-                    { nameof(InitialOffsetOptions.EnqueuedTimeUTC), InitialOffsetOptions.EnqueuedTimeUTC },
+                    { nameof(InitialOffsetOptions.Type), InitialOffsetOptions.Type.ToString() },
+                    { nameof(InitialOffsetOptions.EnqueuedTimeUtc), InitialOffsetOptions.EnqueuedTimeUtc },
                 };
     }
 }