+++ sdk/tables/Azure.Data.Tables/tests/TableEntityTests.cs
@@ -11,7 +11,7 @@ namespace Azure.Tables.Tests
     {
         private TableEntity emptyEntity = new TableEntity { { "My value was nulled!", null } };
         private TableEntity fullEntity = new TableEntity("partition", "row") {
-            { TableConstants.PropertyNames.TimeStamp, default(DateTimeOffset) },
+            { TableConstants.PropertyNames.Timestamp, default(DateTimeOffset) },
             { "binary", new byte[] { 1, 2 }},
             { "boolean", true },
             { "datetime", new DateTime() },